plpgsqltree: move visitor implementations
This is a mechanical commit that moves the `SQLStmtVisitor` and
`TypeRefVisitor` to the `plpgsqltree` package. This will allow the
PL/pgSQL statements in that package to make use of the visitors.

Epic: None

Release note: None
