+++ pkg/sql/catalog/rewrite/BUILD.bazel
@@ -24,7 +24,6 @@ go_library(
         "//pkg/sql/schemachanger/screl",
         "//pkg/sql/sem/catid",
         "//pkg/sql/sem/plpgsqltree",
-        "//pkg/sql/sem/plpgsqltree/utils",
         "//pkg/sql/sem/tree",
         "//pkg/sql/types",
         "//pkg/util/hlc",
