+++ pkg/sql/sem/plpgsqltree/utils/BUILD.bazel
@@ -8,8 +8,6 @@ go_library(
     deps = [
         "//pkg/sql/plpgsql/parser:plpgparser",
         "//pkg/sql/sem/plpgsqltree",
-        "//pkg/sql/sem/tree",
         "//pkg/sql/sqltelemetry",
-        "//pkg/util/errorutil/unimplemented",
     ],
 )
